requirements.txt
fastapi
numpy
requests
Pillow
uvicorn
dwave-systemaltgraph==0.17.3
anyio==3.6.2
appnope==0.1.2
argon2-cffi==21.1.0
attrs==21.2.0
backcall==0.2.0
bleach==4.1.0
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.1.1
click==8.1.3
click-log==0.4.0
colorama==0.4.6
coqpit==0.0.9
coqui-stt-model-manager==0.0.21
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
deprecation==2.1.0
dimod==0.12.4
diskcache==5.4.0
docutils==0.19
dotty-dict==1.3.1
dwave-cloud-client==0.10.4
dwave-networkx==0.8.13
dwave-preprocessing==0.5.4
dwave-samplers==1.0.0
dwave-system==1.18.0
entrypoints==0.3
fastapi==0.94.0
fasteners==0.18
filelock==3.8.0
Flask==2.0.1
Flask-Cors==3.0.10
Flask-SocketIO==4.3.2
gensim==4.2.0
gitdb==4.0.10
GitPython==3.1.31
gotrue==1.0.0
h11==0.14.0
homebase==1.0.1
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.10.0
idna==2.10
importlib-metadata==6.1.0
invoke==1.7.3
ipykernel==6.4.2
ipython==7.29.0
ipython-genutils==0.2.0
itsdangerous==2.1.2
jaraco.classes==3.2.3
jedi==0.18.0
Jinja2==3.0.1
joblib==1.2.0
jsonschema==4.2.1
jupyter-client==7.0.6
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
keyring==23.13.1
macholib==1.16.2
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
minorminer==0.2.11
mistune==0.8.4
more-itertools==9.1.0
mutagen==1.46.0
nbclient==0.5.5
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.8.8
nltk==3.7
notebook==6.4.5
numpy==1.24.2
packaging==21.2
pandocfilters==1.5.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pkginfo==1.9.6
plucky==0.4.3
postgrest==0.10.6
prometheus-client==0.12.0
prompt-toolkit==3.0.21
protobuf==3.20.0
ptyprocess==0.7.0
pycparser==2.21
pycryptodomex==3.16.0
pydantic==1.10.6
Pygments==2.10.0
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.14
pyparsing==2.4.7
PyQt3D==5.15.5
PyQt5==5.15.7
PyQt5-sip==12.11.0
PyQtChart==5.15.6
PyQtDataVisualization==5.15.5
PyQtNetworkAuth==5.15.5
PyQtPurchasing==5.15.5
PyQtWebEngine==5.15.6
pyrsistent==0.18.0
PySocks==1.7.1
python-crfsuite==0.9.9
python-dateutil==2.8.2
python-engineio==3.14.2
python-gitlab==3.13.0
python-semantic-release==7.33.2
python-socketio==4.6.1
pytz==2022.7.1
pyxdg==0.27
PyYAML==6.0
pyzmq==22.3.0
readme-renderer==37.3
realtime==1.0.0
rectangle-packer==2.0.1
regex==2022.9.13
requests==2.28.2
requests-toolbelt==0.10.1
rfc3986==1.5.0
scipy==1.10.1
semver==2.13.0
Send2Trash==1.8.0
sentencepiece==0.1.97
six==1.16.0
smart-open==6.2.0
smmap==5.0.0
sniffio==1.3.0
starlette==0.26.1
storage3==0.5.2
StrEnum==0.4.10
stt==1.4.0
supabase==1.0.2
supafunc==0.2.2
terminado==0.12.1
testpath==0.5.0
tokenizers==0.13.0
tomlkit==0.11.6
torch==1.12.1
tornado==6.1
tqdm==4.64.1
traitlets==5.1.1
transformers @ git+https://github.com/huggingface/transformers@64998a57fbb4d6069c1ce76d7f4c1bf0cd348220
twine==3.8.0
typing_extensions==4.3.0
unidic-lite==1.0.8
urllib3==1.26.12
uvicorn==0.21.0
wcwidth==0.2.5
webencodings==0.5.1
webrtcvad==2.0.10
websockets==10.4
Werkzeug==2.0.3
yt-dlp==2023.2.17
zipp==3.15.0
